named-entity-recognition+image-classification/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.0.0
inflect>=5.0.0
tqdm>=4.0.0
numpy>=1.18.0
scikit-learn>=0.24.0
Pillow>=8.0.0
matplotlib>=3.0.0
seaborn>=0.11.0
nltk>=3.6.0
pandas>=1.0.0
wikipedia-api>=0.5.4